be-rag-to-gcp/requirements.txt
fastapi
python-dotenv
uvicorn
langchain
openai
langchain-openai
langchain-community
PyPDF2
pymupdf
pytesseract
pillow
faiss-cpu
